backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
google-generativeai